+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [1.7.4] - Unreleased
 ### Added
+- GHORG_CLONE_WIKI to clone wiki pages of repos; thanks @ahmadalli
 ### Changed
 ### Deprecated
 ### Removed